+++ CHANGELOG.md
@@ -5,4 +5,5 @@
 - Demo
 
 ### Fixed
+- Fixed issue where deleting the Matlab terminal prevented the extension from running
 - Fixed issue where paths containing spaces were not working.